refactor(timelapse): refactor the timelapse status panel
